requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
annotated-types=0.6.0=py312h06a4308_0
blas=1.0=mkl
brotli-python=1.0.9=py312h6a678d5_9
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.1.31=py312h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
cuda-cudart=12.4.127=h99ab3db_0
cuda-cudart_linux-64=12.4.127=hd681fbe_0
cuda-cupti=12.4.127=h6a678d5_1
cuda-libraries=12.4.1=h06a4308_1
cuda-nvrtc=12.4.127=h99ab3db_1
cuda-nvtx=12.4.127=h6a678d5_1
cuda-opencl=12.4.127=h6a678d5_0
cuda-runtime=12.4.1=hb982923_0
cuda-version=12.4=hbda6634_3
expat=2.6.4=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.17.0=py312h06a4308_0
freetype=2.13.3=h4a9f257_0
giflib=5.2.2=h5eee18b_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py312h5eee18b_0
gnutls=3.6.15=he1e5248_0
idna=3.7=py312h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.6=py312h06a4308_0
jpeg=9e=h5eee18b_3
lame=3.100=h7b6447c_0
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libcublas=12.4.5.8=h99ab3db_1
libcufft=11.2.1.3=h99ab3db_1
libcufile=1.9.1.3=h99ab3db_1
libcurand=10.3.5.147=h99ab3db_1
libcusolver=11.6.1.9=h99ab3db_1
libcusparse=12.3.1.170=h99ab3db_1
libdeflate=1.22=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.2.5.30=h99ab3db_1
libnvfatbin=12.4.127=h7934f7d_2
libnvjitlink=12.4.127=h99ab3db_1
libnvjpeg=12.3.1.117=h6a678d5_1
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=hffd6297_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=3.0.2=py312h5eee18b_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_2
mkl_fft=1.3.11=py312h5eee18b_0
mkl_random=1.2.8=py312h526ad5a_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py312h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.4.2=py312h06a4308_0
numpy=2.0.1=py312hc5e2394_1
numpy-base=2.0.1=py312h0da6c21_1
ocl-icd=2.3.2=h5eee18b_1
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.0.16=h5eee18b_0
pillow=11.1.0=py312hcea889d_0
pip=25.0=py312h06a4308_0
pydantic=2.10.3=py312h06a4308_0
pydantic-core=2.27.1=py312h4aa5aa6_0
pysocks=1.7.1=py312h06a4308_0
python=3.12.9=h5148396_0
pytorch=2.5.1=py3.12_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_7
pytorch-mutex=1.0=cuda
pyyaml=6.0.2=py312h5eee18b_0
readline=8.2=h5eee18b_0
requests=2.32.3=py312h06a4308_1
setuptools=75.8.0=py312h06a4308_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.3=py312h06a4308_1
tbb=2021.8.0=hdb19cb5_0
tk=8.6.14=h39e8969_0
torchtriton=3.1.0=py312
torchvision=0.20.1=py312_cu124
typing-extensions=4.12.2=py312h06a4308_0
typing_extensions=4.12.2=py312h06a4308_0
tzdata=2025a=h04d1e81_0
urllib3=2.3.0=py312h06a4308_0
wheel=0.45.1=py312h06a4308_0
xz=5.6.4=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
